requirements.txt
streamlit>=1.30.0
google-generativeai>=0.5.0
python-dotenv>=1.0.0
pypdf>=4.0.0
langchain>=0.1.0
langchain-google-genai>=1.0.0
langchain-community>=0.2.0
chromadb>=0.5.0
